+++ src/modules/platform/platform.enums.ts
@@ -1,4 +1,5 @@
 export enum PlatformOS {
   Darwin = 'darwin',
   Windows = 'win32',
+  Linux = 'linux',
 }